--- .npmrc
@@ -1,5 +1,0 @@
-link-workspace-packages=true
-public-hoist-pattern[]=*import-in-the-middle*
-public-hoist-pattern[]=*require-in-the-middle*
-public-hoist-pattern[]=*@libsql/client*
-public-hoist-pattern[]=*prettier*
