+++ apps/desktop/package.json
@@ -25,6 +25,7 @@
     "@tauri-apps/plugin-opener": "2.2.6",
     "@trpc/client": "catalog:",
     "@trpc/tanstack-react-query": "catalog:",
+    "hono": "catalog:",
     "react": "catalog:",
     "react-dom": "catalog:",
     "superjson": "catalog:"
