+++ apps/mobile/package.json
@@ -47,11 +47,13 @@
     "expo-haptics": "~14.1.4",
     "expo-linking": "~7.1.5",
     "expo-router": "~5.0.7",
+    "expo-secure-store": "^14.2.3",
     "expo-splash-screen": "~0.30.8",
     "expo-status-bar": "~2.2.3",
     "expo-symbols": "~0.4.4",
     "expo-system-ui": "~5.0.7",
     "expo-web-browser": "~14.1.6",
+    "hono": "catalog:",
     "nativewind": "4.1.23",
     "react": "catalog:",
     "react-dom": "catalog:",
