+++ packages/auth/src/expo/client.ts
@@ -1,0 +1,1 @@
+export { expoClient } from "@better-auth/expo/client"
