+++ packages/db/src/schema/index.ts
@@ -1,6 +1,3 @@
-export * from "./auth"
+export * from "./base"
 export * from "./core"
-export * from "./enums"
-export * from "./organizations"
 export * from "./relations"
-export * from "./sessions"
