+++ packages/utils/package.json
@@ -6,10 +6,7 @@
   "type": "module",
   "exports": {
     ".": "./src/index.ts",
-    "./*": [
-      "./src/helpers/*.ts",
-      "./src/helpers/*.tsx"
-    ],
+    "./*": ["./src/helpers/*.ts", "./src/helpers/*.tsx"],
     "./constants": "./src/constants/index.ts"
   },
   "scripts": {
@@ -23,7 +20,6 @@
   "dependencies": {
     "clsx": "2.1.1",
     "defu": "^6.1.4",
-    "hono": "catalog:",
     "nanoid": "5.1.5",
     "schema-dts": "1.1.5",
     "tailwind-merge": "3.3.0",
