--- packages/utils/src/helpers/hono-client.ts
@@ -1,1 +1,0 @@
-export { hc as createClient } from "hono/client"
